requirements.txt
# Core LangChain and LangGraph
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.8
langgraph>=0.0.40
langsmith>=0.0.83

# Web Interface
streamlit>=1.29.0
streamlit-chat>=0.1.1

# Web Search and APIs
requests>=2.31.0
beautifulsoup4>=4.12.2
google-search-results>=2.4.2
serpapi>=0.1.5

# Image Generation
openai>=1.3.0
pillow>=10.0.0
requests-oauthlib>=1.3.1

# SEO and Content Analysis
textstat>=0.7.3
yake>=0.4.8
readability>=0.3.1

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Utilities
pydantic>=2.5.0
typing-extensions>=4.8.0
python-dateutil>=2.8.2

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional: Database (if you want to store results)
sqlite3
